Refactor chat special buttons
