+++ src/components/chats/ChatSpecialButtons.tsx
@@ -1,0 +1,66 @@
+import AddIcon from '@/assets/icons/add.png'
+import IntegrateIcon from '@/assets/icons/integrate.png'
+import { useSendEvent } from '@/stores/analytics'
+import { cx } from '@/utils/class-names'
+import Image from 'next/image'
+import ChatPreview from './ChatPreview'
+
+export type ChatSpecialButtonsProps = {
+  isIntegrateChatButtonOnTop?: boolean
+}
+
+export default function ChatSpecialButtons({
+  isIntegrateChatButtonOnTop,
+}: ChatSpecialButtonsProps) {
+  const sendEvent = useSendEvent()
+
+  const integrateChatButton = (
+    <ChatPreview
+      key='integrate-chat'
+      isPinned
+      asLink={{ href: '/integrate-chat' }}
+      asContainer
+      onClick={() => sendEvent('click integrate_chat_button')}
+      image={
+        <div className='h-full w-full bg-background-primary p-3 text-text-on-primary'>
+          <Image src={IntegrateIcon} alt='integrate chat' />
+        </div>
+      }
+      className={cx(
+        'bg-background-light md:rounded-none md:bg-background-light/50',
+        isIntegrateChatButtonOnTop ? '' : 'md:rounded-b-3xl'
+      )}
+      withBorderBottom={isIntegrateChatButtonOnTop}
+      title='Integrate chat into an existing app'
+      description='Let your users communicate using blockchain'
+    />
+  )
+
+  const launchCommunityButton = (
+    <ChatPreview
+      key='launch-community'
+      isPinned
+      asLink={{ href: '/launch-community' }}
+      asContainer
+      onClick={() => sendEvent('click launch_community_button')}
+      image={
+        <div className='h-full w-full bg-background-primary p-4 text-text-on-primary'>
+          <Image src={AddIcon} alt='launch community' />
+        </div>
+      }
+      className={cx(
+        'bg-background-light md:rounded-none md:bg-background-light/50',
+        isIntegrateChatButtonOnTop ? 'md:rounded-b-3xl' : ''
+      )}
+      withBorderBottom={!isIntegrateChatButtonOnTop}
+      title='Launch your community'
+      description='Create your own discussion groups'
+    />
+  )
+
+  const specialButtons = isIntegrateChatButtonOnTop
+    ? [integrateChatButton, launchCommunityButton]
+    : [launchCommunityButton, integrateChatButton]
+
+  return <>{specialButtons}</>
+}
