+++ src/modules/chat/HomePage/HomePage.tsx
@@ -1,6 +1,5 @@
-import AddIcon from '@/assets/icons/add.png'
-import IntegrateIcon from '@/assets/icons/integrate.png'
 import ChatPreview from '@/components/chats/ChatPreview'
+import ChatSpecialButtons from '@/components/chats/ChatSpecialButtons'
 import DefaultLayout from '@/components/layouts/DefaultLayout'
 import { getLinkedChatIdsForSpaceId } from '@/constants/chat-room'
 import useIsInIframe from '@/hooks/useIsInIframe'
@@ -8,13 +7,11 @@ import useSearch from '@/hooks/useSearch'
 import { getPostQuery } from '@/services/api/query'
 import { getChatIdsBySpaceIdQuery } from '@/services/subsocial/posts'
 import { useSendEvent } from '@/stores/analytics'
-import { cx } from '@/utils/class-names'
 import { getIpfsContentUrl } from '@/utils/ipfs'
 import { getChatPageLink } from '@/utils/links'
 import { createSlug } from '@/utils/slug'
 import { PostData } from '@subsocial/api/types'
 import dynamic from 'next/dynamic'
-import Image from 'next/image'
 import { useRouter } from 'next/router'
 import { useMemo } from 'react'
 import { useHotkeys } from 'react-hotkeys-hook'
@@ -55,56 +52,6 @@ export default function HomePage({
     searchKeys
   )
 
-  const sendEvent = useSendEvent()
-
-  const integrateChatButton = (
-    <ChatPreview
-      key='integrate-chat'
-      isPinned
-      asLink={{ href: '/integrate-chat' }}
-      asContainer
-      onClick={() => sendEvent('click integrate_chat_button')}
-      image={
-        <div className='h-full w-full bg-background-primary p-3 text-text-on-primary'>
-          <Image src={IntegrateIcon} alt='integrate chat' />
-        </div>
-      }
-      className={cx(
-        'bg-background-light md:rounded-none md:bg-background-light/50',
-        isIntegrateChatButtonOnTop ? '' : 'md:rounded-b-3xl'
-      )}
-      withBorderBottom={isIntegrateChatButtonOnTop}
-      title='Integrate chat into an existing app'
-      description='Let your users communicate using blockchain'
-    />
-  )
-
-  const launchCommunityButton = (
-    <ChatPreview
-      key='launch-community'
-      isPinned
-      asLink={{ href: '/launch-community' }}
-      asContainer
-      onClick={() => sendEvent('click launch_community_button')}
-      image={
-        <div className='h-full w-full bg-background-primary p-4 text-text-on-primary'>
-          <Image src={AddIcon} alt='launch community' />
-        </div>
-      }
-      className={cx(
-        'bg-background-light md:rounded-none md:bg-background-light/50',
-        isIntegrateChatButtonOnTop ? 'md:rounded-b-3xl' : ''
-      )}
-      withBorderBottom={!isIntegrateChatButtonOnTop}
-      title='Launch your community'
-      description='Create your own discussion groups'
-    />
-  )
-
-  const specialButtons = isIntegrateChatButtonOnTop
-    ? [integrateChatButton, launchCommunityButton]
-    : [launchCommunityButton, integrateChatButton]
-
   return (
     <DefaultLayout
       navbarProps={{
@@ -128,7 +75,7 @@ export default function HomePage({
     >
       {!isInIframe && <WelcomeModal />}
       <div className='flex flex-col overflow-auto'>
-        {!isInIframe && !search && specialButtons}
+        {!isInIframe && !search && <ChatSpecialButtons />}
         {searchResults.map((chat, idx) => {
           if (!chat) return null
           return (
