+++ src/modules/chat/HubsPage/HubsPage.tsx
@@ -1,6 +1,5 @@
-import AddIcon from '@/assets/icons/add.png'
-import IntegrateIcon from '@/assets/icons/integrate.png'
 import ChatPreview from '@/components/chats/ChatPreview'
+import ChatSpecialButtons from '@/components/chats/ChatSpecialButtons'
 import DefaultLayout from '@/components/layouts/DefaultLayout'
 import NavbarWithSearch from '@/components/navbar/Navbar/custom/NavbarWithSearch'
 import { getAliasFromSpaceId } from '@/constants/chat-room'
@@ -8,12 +7,10 @@ import useIsInIframe from '@/hooks/useIsInIframe'
 import useSearch from '@/hooks/useSearch'
 import { getSpaceBySpaceIdQuery } from '@/services/subsocial/spaces'
 import { useSendEvent } from '@/stores/analytics'
-import { cx } from '@/utils/class-names'
 import { getSpaceIds } from '@/utils/env/client'
 import { getIpfsContentUrl } from '@/utils/ipfs'
 import { SpaceData } from '@subsocial/api/types'
 import dynamic from 'next/dynamic'
-import Image from 'next/image'
 import { useRouter } from 'next/router'
 import { useMemo } from 'react'
 import { useHotkeys } from 'react-hotkeys-hook'
@@ -45,56 +42,6 @@ export default function HubsPage({
     searchKeys
   )
 
-  const sendEvent = useSendEvent()
-
-  const integrateChatButton = (
-    <ChatPreview
-      key='integrate-chat'
-      isPinned
-      asLink={{ href: '/integrate-chat' }}
-      asContainer
-      onClick={() => sendEvent('click integrate_chat_button')}
-      image={
-        <div className='h-full w-full bg-background-primary p-3 text-text-on-primary'>
-          <Image src={IntegrateIcon} alt='integrate chat' />
-        </div>
-      }
-      className={cx(
-        'bg-background-light md:rounded-none md:bg-background-light/50',
-        isIntegrateChatButtonOnTop ? '' : 'md:rounded-b-3xl'
-      )}
-      withBorderBottom={isIntegrateChatButtonOnTop}
-      title='Integrate chat into an existing app'
-      description='Let your users communicate using blockchain'
-    />
-  )
-
-  const launchCommunityButton = (
-    <ChatPreview
-      key='launch-community'
-      isPinned
-      asLink={{ href: '/launch-community' }}
-      asContainer
-      onClick={() => sendEvent('click launch_community_button')}
-      image={
-        <div className='h-full w-full bg-background-primary p-4 text-text-on-primary'>
-          <Image src={AddIcon} alt='launch community' />
-        </div>
-      }
-      className={cx(
-        'bg-background-light md:rounded-none md:bg-background-light/50',
-        isIntegrateChatButtonOnTop ? 'md:rounded-b-3xl' : ''
-      )}
-      withBorderBottom={!isIntegrateChatButtonOnTop}
-      title='Launch your community'
-      description='Create your own discussion groups'
-    />
-  )
-
-  const specialButtons = isIntegrateChatButtonOnTop
-    ? [integrateChatButton, launchCommunityButton]
-    : [launchCommunityButton, integrateChatButton]
-
   return (
     <DefaultLayout
       navbarProps={{
@@ -123,7 +70,7 @@ export default function HubsPage({
     >
       {!isInIframe && <WelcomeModal />}
       <div className='flex flex-col overflow-auto'>
-        {!isInIframe && !search && specialButtons}
+        {!isInIframe && !search && <ChatSpecialButtons />}
         {searchResults.map((hub, idx) => {
           if (!hub) return null
           return (
